requirements.txt
Flask==1.1.2
gunicorn==20.0.4
robotframework==3.2.2
robotframework-requests==0.7.1
DateTime==4.3
robotframework-sudslibrary-aljcalandra==1.1
urllib3==1.25.10
httplib2==0.18.1
urlparse2==1.1.1
WebTest==2.0.35
jsonpointer==2.0
jsonpatch==1.26
six==1.15.0
webapp2==2.5.2
google-cloud-storage==1.31.0
google==3.0.0
gcloud==0.18.3
webexteamssdk==1.6
regex==2020.7.14
HTMLParser==0.0.2
beautifulsoup4==4.9.1
google-cloud==0.34.0
google-cloud-pubsub==2.1.0
robotframework-selenium2library==3.0.0
google-cloud-logging==2.0.0

